requirements.txt
openai
python-dotenv
requests
pyodbc
streamlit
streamlit_chat
streamlit-scrollable-textbox